% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agitation_tests}
\alias{agitation_tests}
\title{Results of the behavioural test with agitated ants}
\format{
The data frame consists of the following columns:
\describe{
\item{colony}{Colony ID}
\item{date}{Date of the test}
\item{treatment}{Species and number of workers of alien species introduced into the arena}
\item{start}{Time of the test start}
\item{aggression_start}{Experiment start time}
\item{aggression_end}{Experiment end time}
\item{actual_fus_numb}{The number of F. fusca ants on the arena at a given time point; only
ants not involved in aggression were included}
\item{biting_ants_number}{The number of F. fusca ants biting an alien ant on experiment completion}
\item{territory}{Variable indicating location of F. fusca colony relative to red wood ant
territory; T - onto, B - beyond}
\item{aggression_duration}{Duration of alien ant being bitten by F. fusca (in sec.)}
\item{time}{Time and date of experiment start}
\item{alien_colony_ID}{ID of the colony from which alien ants introduced into arena were collected}
}
}
\usage{
agitation_tests
}
\description{
A data frame with the regults of the behavioural tests in which alien ant was placed
into the arena with numerous already agitated F. fusca ants
}
\keyword{datasets}
