% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chi_sq_test}
\alias{chi_sq_test}
\title{Perform the chi-square test for the number of colonies in which
the number of aggressive F. fusca was greater then threshold value (defaults to 0);
the tested factor is location of F. fusca colony relative to red wood ant territory}
\usage{
chi_sq_test(
  upper_limit_date = "2020-12-31",
  lower_limit_date = "2017-01-01",
  treatment_ = "sanguinea_1",
  duration = 60,
  threshold_number = 0,
  raw_data = NULL,
  ...
)
}
\arguments{
\item{upper_limit_date}{A character to be coerced to class 'Date' indicating
upper bound to filter experiments by completion date}

\item{lower_limit_date}{A character to be coerced to class 'Date' indicating
lower bound to filter experiments by completion date}

\item{treatment_}{A character indicating treatment variant to be chosen for
statistical comparison}

\item{duration}{A numeric indicating the number of minutes from the experiment begin
to be considered when determining aggression occurrence.}

\item{threshold_number}{A numeric indicating the minimal number of ants being
aggressive above which the test is classified as showing aggression}

\item{raw_data}{A data frame with source data for the analysis}

\item{...}{Other arguments passed to \code{transform_to_single_val} function}
}
\value{
Output from stats::chisq.test and medians of the compared variables
}
\description{
Perform the chi-square test for the number of colonies in which
the number of aggressive F. fusca was greater then threshold value (defaults to 0);
the tested factor is location of F. fusca colony relative to red wood ant territory
}
