% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPS_coordinates}
\alias{GPS_coordinates}
\title{A data frame with coordinates of the colonies used in the study}
\format{
The data frame consists of the following columns:
\describe{
\item{secies} {ant species}
\item{colony}{ID of s colony}
\item{latitude}{Latitude coordinate of the colony location}
\item{longitude}{Longitude coordinate of the colony location}
}
}
\usage{
GPS_coordinates
}
\description{
A data frame with coordinates of the colonies used in the study
}
\keyword{datasets}
