% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{headwidth_summary}
\alias{headwidth_summary}
\title{Summarize data from measurement of the head width of F. fusca foragers}
\usage{
headwidth_summary(headwidth)
}
\arguments{
\item{headwidth}{A data frame with the results of head width measurement}
}
\value{
A data frame with head width summarized over colonies
}
\description{
Summarize data from measurement of the head width of F. fusca foragers
}
