% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_test_data}
\alias{process_test_data}
\title{Return a data frame row with summarized aggression index for a given test
(maximal number of aggressive ants during selected period);
if range instead of single value is given - minimum value is taken}
\usage{
process_test_data(data, duration = 60, test_id = NA, ...)
}
\arguments{
\item{data}{A data frame with results of the behavioural tests}

\item{duration}{A numeric indicting the period form the test beginning to be
taken into consideration}

\item{test_id}{A numric indicating test ID}

\item{...}{Additional arguments passed to \code{hitherto_number} function}
}
\value{
A data frame with summarized aggression index
}
\description{
Return a data frame row with summarized aggression index for a given test
(maximal number of aggressive ants during selected period);
if range instead of single value is given - minimum value is taken
}
