% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform_to_sigle_val}
\alias{transform_to_sigle_val}
\title{Change data frame by transforming ranges to single values in the specified column}
\usage{
transform_to_sigle_val(
  data = long_tests_results,
  column = "aggression_num",
  mode = 1
)
}
\arguments{
\item{data}{A data frame with source data for the analysis}

\item{column}{A character indicating column name with variable to be transformed}

\item{mode}{A numeric indicating how the range will be mapped to single variable;
1 - minimum, 2 - maximum, 3 - mean, 4 - random.}

\item{treatment}{A character indicating treatment variant to be chosen for
statistical comparison}
}
\value{
A data frame with transformed variable
}
\description{
Change data frame by transforming ranges to single values in the specified column
}
